Переработал интерфейс и функции для лучшего внедрения mvvm
